Introduce new super class RegexImportFactory
