+++ src/main/java/openfasttrack/importer/ImporterFactory.java
@@ -4,7 +4,7 @@ package openfasttrack.importer;
  * #%L
  * OpenFastTrack
  * %%
- * Copyright (C) 2016 - 2017 hamstercommunity
+ * Copyright (C) 2016 - 2018 hamstercommunity
  * %%
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as
@@ -22,20 +22,13 @@ package openfasttrack.importer;
  * #L%
  */
 
-
-import static java.util.Arrays.asList;
-import static java.util.stream.Collectors.toSet;
-
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.Reader;
 import java.nio.charset.Charset;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.Collection;
-import java.util.Set;
 import java.util.logging.Logger;
-import java.util.regex.Pattern;
 
 /**
  * Super class for factories producing {@link Importer}s.
@@ -44,45 +37,15 @@ public abstract class ImporterFactory
 {
     private static final Logger LOG = Logger.getLogger(ImporterFactory.class.getName());
 
-    private final Set<Pattern> supportedFilenamePatterns;
-
-    protected ImporterFactory(final String... supportedFilenamePatterns)
-    {
-        this(asList(supportedFilenamePatterns));
-    }
-
-    protected ImporterFactory(final Collection<String> supportedFileExtensions)
-    {
-        this.supportedFilenamePatterns = supportedFileExtensions.stream() //
-                .map(Pattern::compile) //
-                .collect(toSet());
-    }
-
     /**
-     * Returns <code>true</code> if this {@link ImporterFactory} supports
-     * importing the given file based on its file extension.
+     * Returns <code>true</code> if this {@link ImporterFactory} supports importing
+     * the given file based on its file extension.
      *
      * @param file
      *            the file to check.
      * @return <code>true</code> if the given file is supported for importing.
      */
-    public boolean supportsFile(final Path file)
-    {
-        final String fileName = file.getFileName().toString();
-        for (final Pattern pattern : this.supportedFilenamePatterns)
-        {
-            if (pattern.matcher(fileName).matches())
-            {
-                LOG.finest(() -> "Filename '" + fileName + "' matches '" + pattern
-                        + "': supported  by " + this.getClass().getName());
-                return true;
-            }
-        }
-        LOG.finest(() -> "Filename '" + fileName + "' does not match any regexp of "
-                + this.supportedFilenamePatterns + ": not supported by "
-                + this.getClass().getName());
-        return false;
-    }
+    public abstract boolean supportsFile(final Path file);
 
     /**
      * Create an importer that is able to read the given file.
